Refactor the ToUnstructured functions
Change resource.ToUnstructured to use the
runtime.DefaultUnstructuredConverter instead of
scheme.Scheme.Convert as the latter requres the type
to be registered with the global scheme.

Also move the test ToUnstructured function from the syncer/test
package to the general test package.
